alterações finais ao código para todas as implementações

+++ 93_ES/src/main/java/_ES/_ES/ThresHold.java
@@ -34,7 +34,7 @@ public class ThresHold {
 		return value;
 	}
 
-	public void setValue(int value) {
+	public void setValue(double value) {
 		this.value = value;
 	}
 }
